+++ python/GafferUI/SpreadsheetUI/_EditWindow.py
@@ -57,18 +57,13 @@ class _EditWindow( GafferUI.Window ) :
 		self.keyPressSignal().connect( Gaffer.WeakMethod( self.__keyPress ), scoped = False )
 
 	@classmethod
-	def popupEditor( cls, plug, plugBound ) :
+	def popupEditor( cls, plugs, plugBound ) :
 
-		plugValueWidget = GafferUI.PlugValueWidget.create( plug )
+		if not isinstance( plugs, set ) :
+			plugs = set( plugs )
+
+		plugValueWidget = GafferUI.PlugValueWidget.create( plugs )
 		cls.__currentWindow = _EditWindow( plugValueWidget )
-
-		if isinstance( plugValueWidget, _CellPlugValueWidget ) :
-			valuePlugValueWidget = plugValueWidget.childPlugValueWidget( plug["value"] )
-			if isinstance( valuePlugValueWidget, GafferUI.PresetsPlugValueWidget ) :
-				if not Gaffer.Metadata.value( valuePlugValueWidget.getPlug(), "presetsPlugValueWidget:isCustom" ) :
-					valuePlugValueWidget.menu().popup()
-					return
-
 		cls.__currentWindow.resizeToFitChild()
 		windowSize = cls.__currentWindow.bound().size()
 		cls.__currentWindow.setPosition( plugBound.center() - windowSize / 2 )
@@ -120,7 +115,7 @@ class _EditWindow( GafferUI.Window ) :
 		if widget is not None and widgetUsable( widget ) :
 			return widget
 
-		for childPlug in Gaffer.Plug.Range( plugValueWidget.getPlug() ) :
+		for childPlug in Gaffer.Plug.Range( next( iter( plugValueWidget.getPlugs() ) ) ) :
 			childWidget = plugValueWidget.childPlugValueWidget( childPlug )
 			if childWidget is not None :
 				childTextWidget = cls.__textWidget( childWidget )