+++ Changes.md
@@ -19,6 +19,15 @@ API
 0.58.5.2 (relative to 0.58.5.1)
 ========
 
+Improvements
+------------
+
+- Spreadsheet :
+  - Enabled cell selection, editing is now achieved with a double-click.
+  - Added menu items to cells & defaults sections to copy/paste the enabled state and value(s) of selected cells.
+  - Added menu items to the row names section to copy/paste or delete the selected rows.
+  - Added menu items to edit the values for all selected cells of the same type simultaneously.
+
 Fixes
 -----